+++ src/__stories__/search-field-story.tsx
@@ -20,6 +20,8 @@ interface SearchFieldBaseArgs {
     error: boolean;
     inverse: boolean;
     optional: boolean;
+    showOptionalLabel: boolean;
+    withStartIcon: boolean;
     disabled: boolean;
     readOnly: boolean;
     preventCopy: boolean;
@@ -32,6 +34,8 @@ const defaultBaseArgs: SearchFieldBaseArgs = {
     error: false,
     inverse: false,
     optional: false,
+    showOptionalLabel: true,
+    withStartIcon: true,
     disabled: false,
     readOnly: false,
     preventCopy: false,