+++ src/fixed-footer-layout.tsx
@@ -163,6 +163,13 @@ const FixedFooterLayout = ({
                 {renderBackground()}
                 {children}
             </div>
+            {isFooterVisible && isFooterFixed && (
+                /**
+                 * This style is added to the `html` element to prevent the footer
+                 * from covering the focus of elements when there is overflow.
+                 */
+                <style>{`html {scroll-padding-bottom: ${domFooterHeight}px}`}</style>
+            )}
             <div
                 className={classnames(styles.footer, {
                     [styles.withoutFooter]: !isFooterVisible,